ok - this a step backwards before we go forwards
this removes the lite version of the sdk and also the full one.
It moves the old lite code into the common directory and removes the lite directory.

I've not tried building the common sdk yet because I'm figuring out the way to make the protobuf stuff.

Here's my planned folder structure:

build
    - grpc_service
      - preprocessed - the preprocessed proto files
      - generated - the generated c++ headers and body files
      libgrpc_service.a

    - proto_interface_lite
        - preprocessed - the preprocessed proto files
        - generated - the generated c++ headers and body files
        libproto_interface_lite.a

    - proto_interface_full
        - preprocessed - the preprocessed proto files
        - generated - the generated c++ headers and body files
        libproto_interface_full.a

the grpc_service is built if gRPC is specified as a CONNECTIONS

the proto_interface_lite is built come what may - the common SDK will link against It

the proto_interface_full is built if REST is specified as a CONNECTIONS
