+++ sdks/cpp/common/include/NamedChoiceConstraint.h
@@ -1,21 +1,34 @@
 #pragma once
 
-/** Copyright 2024 Ross Video Ltd
-
- Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
-
- 1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer.
-
- 2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer in the documentation and/or other materials provided with the distribution.
-
- 3. Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote products derived from this software without specific prior written permission.
-
- THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS “AS IS” AND ANY EXPRESS OR IMPLIED WARRANTIES, 
- INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, 
- INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER 
- CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*/
-//
+/*
+ * Copyright 2024 Ross Video Ltd
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice,
+ * this list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ *
+ * 3. Neither the name of the copyright holder nor the names of its
+ * contributors may be used to endorse or promote products derived from this
+ * software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS “AS IS”
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * RE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ */
 
 /**
  * @file NamedChoiceConstraint.h
@@ -29,8 +42,6 @@
 #include <IConstraint.h>
 #include <IParam.h>
 #include <Tags.h>
-
-// lite
 #include <PolyglotText.h>
 #include <Device.h>
 
@@ -41,15 +52,14 @@
 #include <initializer_list>
 
 namespace catena {
-namespace lite {
+namespace common {
 
 /**
  * @brief Named choice constraint, ensures a value is within a named choice
  * @tparam T int or string
  */
-template <typename T>
-class NamedChoiceConstraint : public catena::common::IConstraint {
-public:
+template <typename T> class NamedChoiceConstraint : public catena::common::IConstraint {
+  public:
     /**
      * @brief map of choices with their display names
      */
@@ -59,7 +69,7 @@ public:
      */
     using ListInitializer = std::initializer_list<std::pair<T, PolyglotText::ListInitializer>>;
 
-public:
+  public:
     /**
      * @brief Construct a new Named Choice Constraint object
      * @param init the list of choices
@@ -70,11 +80,11 @@ public:
      * @note  the first choice provided will be the default for the constraint
      */
     NamedChoiceConstraint(ListInitializer init, bool strict, std::string oid, bool shared, Device& dm)
-        : choices_{init.begin(), init.end()}, 
-        strict_{strict}, default_{init.begin()->first}, oid_{oid}, shared_{shared} {
+        : choices_{init.begin(), init.end()}, strict_{strict}, default_{init.begin()->first}, oid_{oid},
+          shared_{shared} {
         dm.addItem<common::ConstraintTag>(oid, this);
     }
-    
+
     /**
      * @brief Construct a new Named Choice Constraint object
      * @param init the list of choices
@@ -85,8 +95,8 @@ public:
      * @note  the first choice provided will be the default for the constraint
      */
     NamedChoiceConstraint(ListInitializer init, bool strict, std::string oid, bool shared)
-        : choices_{init.begin(), init.end()}, 
-        strict_{strict}, default_{init.begin()->first}, oid_{oid}, shared_{shared} {}
+        : choices_{init.begin(), init.end()}, strict_{strict}, default_{init.begin()->first}, oid_{oid},
+          shared_{shared} {}
 
     /**
      * @brief default destructor
@@ -100,11 +110,11 @@ public:
      */
     bool satisfied(const catena::Value& src) const override {
 
-        if constexpr(std::is_same<T, int32_t>::value) {
+        if constexpr (std::is_same<T, int32_t>::value) {
             return choices_.find(src.int32_value()) != choices_.end();
         }
 
-        if constexpr(std::is_same<T, std::string>::value) {
+        if constexpr (std::is_same<T, std::string>::value) {
             if (!strict_) {
                 return true;
             }
@@ -116,10 +126,10 @@ public:
      * @brief applies constraint to src and returns the constrained value
      * @param src a catena::Value to apply the constraint to
      * @return an empty catena::Value
-     * 
+     *
      * If a request does not satisfy a choice constraint, then
      * the request is invalid and should be ignored.
-     * 
+     *
      * Calling this will always return an empty value.
      */
     catena::Value apply(const catena::Value& src) const override {
@@ -133,7 +143,7 @@ public:
      */
     void toProto(catena::Constraint& constraint) const override {
 
-        if constexpr(std::is_same<T, int32_t>::value) {
+        if constexpr (std::is_same<T, int32_t>::value) {
             constraint.set_type(catena::Constraint::INT_CHOICE);
             for (auto& [value, name] : choices_) {
                 auto intChoice = constraint.mutable_int32_choice()->add_choices();
@@ -142,7 +152,7 @@ public:
             }
         }
 
-        if constexpr(std::is_same<T, std::string>::value) {
+        if constexpr (std::is_same<T, std::string>::value) {
             constraint.set_type(catena::Constraint::STRING_STRING_CHOICE);
             for (auto& [value, name] : choices_) {
                 auto stringChoice = constraint.mutable_string_string_choice()->add_choices();
@@ -170,13 +180,13 @@ public:
      */
     const std::string& getOid() const override { return oid_; }
 
-private:
-    Choices choices_; ///< the choices
-    bool strict_;     ///< should the value be constrained on apply
-    T default_;       ///< the default value to constrain to
-    bool shared_;     ///< is the constraint shared
-    std::string oid_; ///< the oid of the constraint
+  private:
+    Choices choices_;  ///< the choices
+    bool strict_;      ///< should the value be constrained on apply
+    T default_;        ///< the default value to constrain to
+    bool shared_;      ///< is the constraint shared
+    std::string oid_;  ///< the oid of the constraint
 };
 
-} // namespace lite
-} // namespace catena
+}  // namespace common
+}  // namespace catena
