+++ sdks/cpp/common/include/PicklistConstraint.h
@@ -1,21 +1,34 @@
 #pragma once
 
-/** Copyright 2024 Ross Video Ltd
-
- Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
-
- 1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer.
-
- 2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer in the documentation and/or other materials provided with the distribution.
-
- 3. Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote products derived from this software without specific prior written permission.
-
- THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS “AS IS” AND ANY EXPRESS OR IMPLIED WARRANTIES, 
- INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, 
- INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER 
- CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*/
-//
+/*
+ * Copyright 2024 Ross Video Ltd
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice,
+ * this list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ *
+ * 3. Neither the name of the copyright holder nor the names of its
+ * contributors may be used to endorse or promote products derived from this
+ * software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS “AS IS”
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * RE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ */
 
 /**
  * @file PicklistConstraint.h
@@ -29,8 +42,6 @@
 #include <IConstraint.h>
 #include <IParam.h>
 #include <Tags.h>
-
-// lite
 #include <Device.h>
 
 #include <google/protobuf/message_lite.h>
@@ -40,13 +51,13 @@
 #include <initializer_list>
 
 namespace catena {
-namespace lite {
+namespace common {
 
 /**
  * @brief Picklist constraint, ensures a value is in a list of strings
  */
-class PicklistConstraint : public catena::common::IConstraint {
-public:
+class PicklistConstraint : public IConstraint {
+  public:
     /**
      * @brief map of choices with their display names
      */
@@ -56,7 +67,7 @@ public:
      */
     using ListInitializer = std::initializer_list<std::string>;
 
-public:
+  public:
     /**
      * @brief Construct a new Picklist Constraint object
      * @param init the list of choices
@@ -67,7 +78,7 @@ public:
      * @note  the first choice provided will be the default for the constraint
      */
     PicklistConstraint(ListInitializer init, bool strict, std::string oid, bool shared, Device& dm);
-    
+
     /**
      * @brief Construct a new Picklist Constraint object
      * @param init the list of choices
@@ -91,14 +102,14 @@ public:
      */
     bool satisfied(const catena::Value& src) const override;
 
-     /**
+    /**
      * @brief applies constraint to src and returns the constrained value
      * @param src a catena::Value to apply the constraint to
      * @return an empty catena::value
-     * 
+     *
      * If a request does not satisfy a choice constraint, then
      * the request is invalid and should be ignored.
-     * 
+     *
      * Calling this will always return an empty value.
      */
     catena::Value apply(const catena::Value& src) const override;
@@ -127,13 +138,13 @@ public:
      */
     const std::string& getOid() const override { return oid_; }
 
-private:
-    Choices choices_;     ///< the choices
-    bool strict_;         ///< should the value be constrained on apply
-    std::string default_; ///< the default value to constrain to
-    bool shared_;         ///< is the constraint shared
-    std::string oid_;     ///< the oid of the constraint
+  private:
+    Choices choices_;      ///< the choices
+    bool strict_;          ///< should the value be constrained on apply
+    std::string default_;  ///< the default value to constrain to
+    bool shared_;          ///< is the constraint shared
+    std::string oid_;      ///< the oid of the constraint
 };
 
-} // namespace lite
-} // namespace catena
+}  // namespace common
+}  // namespace catena
