+++ sdks/cpp/common/src/PolyglotText.cpp
@@ -1,5 +1,3 @@
-#pragma once
-
 /*
  * Copyright 2024 Ross Video Ltd
  *
@@ -30,35 +28,16 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-/**
- * @file IPolyglotText.h
- * @brief Catena's multi-language text interface
- * @author John R. Naylor, john.naylor@rossvideo.com
- * @copyright Copyright (c) 2024 Ross Video
- */
+#include <PolyglotText.h>
 
-#include "google/protobuf/message_lite.h" 
-
-#include <unordered_map>
 #include <string>
-#include <initializer_list>
+#include <unordered_map>
 
-namespace catena {
-namespace common {
-class IPolyglotText {
-  public:
-    using DisplayStrings = std::unordered_map<std::string, std::string>;
-    using ListInitializer = std::initializer_list<std::pair<std::string, std::string>>;
-  public:
-    IPolyglotText() = default;
-    IPolyglotText(IPolyglotText&&) = default;
-    IPolyglotText& operator=(IPolyglotText&&) = default;
-    virtual ~IPolyglotText() = default;
-
-    virtual void toProto(google::protobuf::MessageLite& msg) const = 0;
-
-    virtual const DisplayStrings& displayStrings() const = 0;
-
-};
-}  // namespace common
-}  // namespace catena
+using catena::common::PolyglotText;
+void PolyglotText::toProto(google::protobuf::MessageLite& m) const {
+    auto& dst = dynamic_cast<catena::PolyglotText&>(m);
+    dst.clear_display_strings();
+    for (const auto& [key, value] : display_strings_) {
+        dst.mutable_display_strings()->insert({key, value});
+    }
+}
