--- sdks/cpp/lite/src/LanguagePack.cpp
@@ -1,44 +1,0 @@
-/** Copyright 2024 Ross Video Ltd
-
- Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
-
- 1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following disclaimer.
-
- 2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer in the documentation and/or other materials provided with the distribution.
-
- 3. Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote products derived from this software without specific prior written permission.
-
- THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS “AS IS” AND ANY EXPRESS OR IMPLIED WARRANTIES, 
- INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, 
- INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER 
- CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*/
-//
-
-// lite
-#include <LanguagePack.h>
-
-using namespace catena::lite;
-using catena::common::LanguagePackTag;
-
-LanguagePack::LanguagePack(const std::string& name, ListInitializer list, Device& dev)
-    : name_{name}, words_(list.begin(), list.end()) {
-    dev.addItem<LanguagePackTag>(name, this);
-}
-
-void LanguagePack::fromProto(const catena::LanguagePack& pack) {
-    name_ = pack.name();
-    for (const auto& [key, value] : pack.words()) {
-        words_[key] = value;
-    }
-}
-
-void LanguagePack::toProto(catena::LanguagePack& pack) const {
-    pack.set_name(name_);
-    for (const auto& [key, value] : words_) {
-        (*pack.mutable_words())[key] = value;
-    }
-}
-
-
-
